+++ index.html
@@ -231,5 +231,7 @@
   </template>
 
 <script src="./js/main.js"></script>
+<script src="./js/data.js" type="module"></script>
+<script src="./js/util.js" type="module"></script>
 </body>
 </html>
